requirements.txt
Flask
flask-cors
firebase-admin
cloudinary
python-dotenv
waitress 